requirements.txt
fastapi[standard]
uvicorn
phonenumbers
folium
opencage
python-dotenv
pydantic